@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk.Razor">
 
 	<PropertyGroup>
-		<TargetFramework>net5.0</TargetFramework>
+		<TargetFramework>net6.0</TargetFramework>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="5.0.11" />
-		<PackageReference Include="NuGet.Versioning" Version="5.11.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="6.0.0" />
+		<PackageReference Include="NuGet.Versioning" Version="6.0.0" />
 		<PackageReference Include="Prism.Core" Version="8.0.0.1909" />
 	</ItemGroup>
 
