@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using F0.Minesweeper.Logic.Abstractions;
 using F0.Minesweeper.Logic.LocationShuffler;
@@ -14,7 +12,7 @@ namespace F0.Minesweeper.Logic
 
 		public IMinefield Create(MinefieldOptions options)
 		{
-			_ = options ?? throw new ArgumentNullException(nameof(options));
+			ArgumentNullException.ThrowIfNull(options);
 
 			ILocationShuffler locationShuffler;
 
@@ -26,7 +24,9 @@ namespace F0.Minesweeper.Logic
 			{
 				locationShuffler = options.LocationShuffler switch
 				{
-					Abstractions.LocationShuffler.GuidLocationShuffler => new GuidLocationShuffler(),
+					Abstractions.LocationShuffler.GloballyUniqueIdentifier => new GuidLocationShuffler(),
+					Abstractions.LocationShuffler.RandomOrder => new RandomOrderLocationShuffler(),
+					Abstractions.LocationShuffler.FisherYates => new FisherYatesLocationShuffler(),
 					_ => throw new InvalidEnumArgumentException(nameof(options.LocationShuffler), (int)options.LocationShuffler, typeof(Abstractions.LocationShuffler)),
 				};
 
