@@ -2,6 +2,8 @@ namespace F0.Minesweeper.Logic.Abstractions
 {
 	public enum LocationShuffler
 	{
-		GuidLocationShuffler,
+		GloballyUniqueIdentifier,
+		RandomOrder,
+		FisherYates
 	}
 }
